feat: new struct, separate event and config ticket
